バッチ修正

+++ data-linkage/codes/nayose/process/Process35.php
@@ -18,7 +18,7 @@ class Process35{
 	}
 	
 	/**
-	 * Execute Process 19
+	 * Execute Process35
 	 */
 	function execProcess(){
 		try{
@@ -45,40 +45,65 @@ class Process35{
 	 */
 	function renameDir($dir){
 		try{
-			//check if file exist
+			//add 「_comp」 in the end of the name
+			$compFolderePath = rtrim($dir, '/').'_comp';
+			
+			//$this->logger->debug($dir);
 			if(glob($dir)){
-				//add 「_comp」 in the end of the name
-				$compFolderePath = rtrim($dir, '/').'_comp';
-				foreach(glob($dir.'*') as $f_name){
-					//$this->logger->debug($f_name);
-					
-					$this->mkdirBackupFoldere($compFolderePath);
-					/****
-					// ファイル名取得
-					$file_name =shell_exec('basename '.$f_name);
-					// 改行が入るため取る
-					$file_name=str_replace(array("\r\n","\n","\r"), '', $file_name);
-					// ディレクトリ名
-					$dirname = dirname($f_name);
-				
-					shell_exec('tar zcvf '.escapeshellarg($compFolderePath.'/'.$file_name.'.tar.gz').' -C '.escapeshellarg($dirname).' '.escapeshellarg($file_name).' --remove-files');
-					$this->logger->debug('moved and compression '.$f_name.' to '.$compFolderePath);
-					*/
-					shell_exec('mv -f '.escapeshellarg($f_name).' '.escapeshellarg($compFolderePath));
-					$this->logger->debug('moved '.$f_name.' to '.$compFolderePath);
+				foreach(glob($dir.'*/') as $f_name1){
+					if (is_dir($f_name1)) {
+						$d_name = $f_name1;
+						// ディレクトリ名取得
+						$subdir_name = basename($f_name1);
+						
+						foreach(glob($d_name.'*/') as $f_name2){
+							$this->movefile($f_name2, $compFolderePath.'/'.$subdir_name);
+						}
+						$this->rmdirDir($d_name, $compFolderePath);
+					} else {
+						$this->movefile($f_name1, $compFolderePath);
+					}
 				}
+			}
+			$this->rmdirDir($dir, $compFolderePath);
+		}catch(Exception $e){
+			throw $e;
+		}
+	}
+	
+	function movefile($f_name, $compFolderePath){
+		$this->logger->debug($f_name);
+		$this->mkdirBackupFoldere($compFolderePath);
+		/****
+		// ファイル名取得
+		$file_name =shell_exec('basename '.$f_name);
+		// 改行が入るため取る
+		$file_name=str_replace(array("\r\n","\n","\r"), '', $file_name);
+		// ディレクトリ名
+		$dirname = dirname($f_name);
+	
+		shell_exec('tar zcvf '.escapeshellarg($compFolderePath.'/'.$file_name.'.tar.gz').' -C '.escapeshellarg($dirname).' '.escapeshellarg($file_name).' --remove-files');
+		$this->logger->debug('moved and compression '.$f_name.' to '.$compFolderePath);
+		*/
+		shell_exec('mv -f '.escapeshellarg($f_name).' '.escapeshellarg($compFolderePath));
+		$this->logger->debug('moved '.$f_name.' to '.$compFolderePath);
+	}
+	
+	function rmdirDir($dir, $compFolderePath){
+		if(glob($dir)){
+			// ディレクトリ削除
+			if (count(glob($dir.'*/')) == 0) {
 				// ディレクトリ削除
 				if(rmdir($dir)){
 					// 削除成功
-					$this->logger->info('removed '.$dir);
+					$this->logger->debug('removed '.$dir);
 				} else {
 					// 削除失敗
 					$this->logger->error("ディレクトリ削除に失敗しました. folder name：".$dir);
-					throw new Exception("Process19 Failed to remove folders. ".$dir);
+					throw new Exception("Process35 Failed to remove folders. ".$dir);
 				}
+				$this->logger->info('renamed '.$dir.' to '.$compFolderePath);
 			}
-		}catch(Exception $e){
-			throw $e;
 		}
 	}
